requirements.txt
flask==3.0.0
selenium==4.9.0
pytest==7.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest-html==3.2.0